Fix repeat timer calculation
If your lircd.conf has "min_repeat 1" and you call irsend twice in quick
succession (like this: `irsend SEND_ONCE $remote $key; irsend SEND_ONCE
$remote $key`) then this happens:

1. First irsend: lircd sends first signal (0th repeat), then schedules a
   timer to trigger after the gap has elapsed. Logging says:

   > Trace: alarm in 39897 usecs

2. When the timer ends we get SIGALRM and we send the 1st repeat signal.

3. Now the 2nd irsend is received immediately after the previous step.
   lircd wants to send the first signal (0th repeat) but it's too soon
   after `last_send`, so it usleeps for a suitable gap *before* sending
   the signal.

   [Incidentally, why does it sleep for `min_remaining_gap * 2` -- I
   mean why the `* 2`? This was added 22 years ago in Nov 2000 in commit
   952920215 which merely says "better min_repeat handling".]

   Then it sends the first signal (0th repeat) and schedules a timer to
   trigger after the gap has elapsed. Here is the problem: We calculate
   the timer based on the time elapsed since *before* the usleep! Since
   we've already slept for longer than the repeat gap the timer is going
   to fire immediately. Logging says:

   > Trace: alarm in 10 usecs

4. We get SIGALRM and send the 1st repeat signal. Here again we need to
   usleep because we're trying to send it too soon after `last_send`.

So the behaviour used to be: If you do send_once (or send_start, I
suppose) too hot on the heels of a previous signal, during the gaps you
end up blocking in `send_ir_ncode`, instead of waiting in the main loop
in `mywaitfordata`. So the processing of new client connections will be
blocked for the duration of the gap.

The solution: Don't count the time we spent sleeping in step 3, when
we're calculating the timer for sending the next repeat. It turns out
`send_ir_ncode` only has 2 callers with `delay=1` so I can pull that
delay out of `send_ir_ncode` into `send_core`. We don't need to
check/sleep when sending repeat signals (in `dosigalrm`) because we
have already used the timer to wait long enough.

The very short timer (10 usecs) caused by this bug was triggering a race
condition in lircd's sigalrm handling. I'll fix the race condition in a
separate pull request.