Transmit second code rather than repeat header
In IP42RC.lircd.conf we have multiple codes per button, and a repeat header.
After sending the first code, we send the repeat header rather than the second
code. This is not enough to decode. Fix this by not sending the repeat
header if there are more codes to send.

With this change, lircd can decode what it transmits for IP42RC.lircd.conf.